requirements.txt
Cython==3.0.11
datasets==2.13.2
diffusers==0.4.1
grpcio==1.62.3
huggingface-hub==0.16.4
imageio==2.31.2
Jinja2==3.1.4
mmcv-full==1.7.0
mmdet==2.25.3
mmpycocotools==12.0.3
numpy==1.21.6
opencv-python==4.10.0.84
pandas==1.3.5
Pillow==9.5.0
pycocotools==2.0.7
pytz==2023.4
PyYAML==6.0.1
scikit-image==0.19.3
scipy==1.7.3
tensorboard==2.11.2
timm==0.9.12
torch==1.12.1+cu113
torch-fidelity==0.3.0
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.65.2
transformers==4.24.0
wandb==0.17.7